Move Pipeline to a file
