+++ src/sentry/digests/notifications.py
@@ -4,15 +4,15 @@ import functools
 import itertools
 import logging
 from collections import OrderedDict, defaultdict, namedtuple
-from functools import reduce
-from typing import Any, Callable, Mapping, MutableMapping, MutableSequence, Sequence
+from typing import Any, Mapping, MutableMapping, MutableSequence, Sequence
 
 from sentry.app import tsdb
-from sentry.digests import Record
+from sentry.digests import Digest, Record
 from sentry.eventstore.models import Event
 from sentry.models import Group, GroupStatus, Project, Rule
 from sentry.notifications.types import ActionTargetType
 from sentry.utils.dates import to_timestamp
+from sentry.utils.pipeline import Pipeline
 
 logger = logging.getLogger("sentry.digests")
 
@@ -101,59 +101,6 @@ def attach_state(
     return {"project": project, "groups": groups, "rules": rules}
 
 
-class Pipeline:
-    def __init__(self) -> None:
-        self.operations: MutableSequence[Callable[..., Any]] = []
-        self.logs: MutableSequence[str] = []
-
-    def __call__(self, sequence: Sequence[Any]) -> tuple[Any, Sequence[str]]:
-        # Explicitly typing to satisfy mypy.
-        func: Callable[[Any, Callable[[Any], Any]], Any] = lambda x, operation: operation(x)
-        return reduce(func, self.operations, sequence), self.logs
-
-    def _log(self, message: str) -> None:
-        logger.debug(message)
-        self.logs.append(message)
-
-    def apply(self, function: Callable[[MutableMapping[str, Any]], Any]) -> Pipeline:
-        def operation(sequence: MutableMapping[str, Any]) -> Any:
-            result = function(sequence)
-            self._log(f"{function!r} applied to {len(sequence)} items.")
-            return result
-
-        self.operations.append(operation)
-        return self
-
-    def filter(self, function: Callable[[Record], bool]) -> Pipeline:
-        def operation(sequence: Sequence[Any]) -> Sequence[Any]:
-            result = [s for s in sequence if function(s)]
-            self._log(f"{function!r} filtered {len(sequence)} items to {len(result)}.")
-            return result
-
-        self.operations.append(operation)
-        return self
-
-    def map(self, function: Callable[[Sequence[Any]], Any]) -> Pipeline:
-        def operation(sequence: Sequence[Any]) -> Sequence[Any]:
-            result = [function(s) for s in sequence]
-            self._log(f"{function!r} applied to {len(sequence)} items.")
-            return result
-
-        self.operations.append(operation)
-        return self
-
-    def reduce(
-        self, function: Callable[[Any, Any], Any], initializer: Callable[[Sequence[Any]], Any]
-    ) -> Pipeline:
-        def operation(sequence: Sequence[Any]) -> Any:
-            result = reduce(function, sequence, initializer(sequence))
-            self._log(f"{function!r} reduced {len(sequence)} items to {len(result)}.")
-            return result
-
-        self.operations.append(operation)
-        return self
-
-
 def rewrite_record(
     record: Record,
     project: Project,
@@ -217,30 +164,26 @@ def sort_rule_groups(rules: Mapping[str, Rule]) -> Mapping[str, Rule]:
     )
 
 
+def check_group_state(record: Record) -> bool:
+    # Explicitly typing to satisfy mypy.
+    is_unresolved: bool = record.value.event.group.get_status() == GroupStatus.UNRESOLVED
+    return is_unresolved
+
+
 def build_digest(
     project: Project,
     records: Sequence[Record],
     state: Mapping[str, Any] | None = None,
-) -> tuple[Any | None, Sequence[str]]:
-    records = list(records)
+) -> tuple[Digest | None, Sequence[str]]:
     if not records:
         return None, []
 
-    # XXX: This is a hack to allow generating a mock digest without actually
-    # doing any real IO!
-    if state is None:
-        state = fetch_state(project, records)
-
-    state = attach_state(**state)
-
-    def check_group_state(record: Record) -> bool:
-        # Explicitly typing to satisfy mypy.
-        is_unresolved: bool = record.value.event.group.get_status() == GroupStatus.UNRESOLVED
-        return is_unresolved
+    # XXX(hack): Allow generating a mock digest without actually doing any real IO!
+    state = state or fetch_state(project, records)
 
     pipeline = (
         Pipeline()
-        .map(functools.partial(rewrite_record, **state))
+        .map(functools.partial(rewrite_record, **attach_state(**state)))
         .filter(bool)
         .filter(check_group_state)
         .reduce(group_records, lambda sequence: defaultdict(lambda: defaultdict(list)))
@@ -248,4 +191,5 @@ def build_digest(
         .apply(sort_rule_groups)
     )
 
-    return pipeline(records)
+    digest, logs = pipeline(records)
+    return digest, logs
